🔨 Moved discord rich presence handling to its own module
