+++ src-tauri/src/discord/mod.rs
@@ -1,0 +1,2 @@
+pub mod drpc;
+pub use drpc::*;
